perf: ape test
